notice 페이지 완성
