+++ components/employee/index.ts
@@ -1,9 +1,11 @@
 import EditProfile from "./EditProfile/EditProfile";
 import EmployeeNotice from "./EmployeeNotice/EmployeeNotice";
 import RegisteredMyProfile from "./RegisteredMyProfile/RegisteredMyProfile";
+import ApplicationDetails from "./ApplicationDetails/ApplicationDetails";
 
 export {
   EditProfile,
   EmployeeNotice,
   RegisteredMyProfile,
+  ApplicationDetails,
 };
